Move Application class to its own file
